{"schema":"decodex/commit/1","summary":"Extract live HUD color roll coordinator","authority":"manual"}
