+++ docs/reference/host-core-reset.md
@@ -125,6 +125,9 @@ The current native-host Swift split is:
   with per-track throttling state, and AppKit-to-controller pointer delivery support.
 - `LiveOverlayRenderer.swift`: live overlay render orchestration for HUD, loupe, frozen-pending
   preview, and live chrome layers.
+- `LiveHudColorRollCoordinator.swift`: live HUD color swatch/hex presentation, pending color roll
+  animation state, resolved hex roll transitions, and roll-layer lifecycle.
+- `LiveOverlayTypography.swift`: shared native live overlay font and text metrics.
 - `LiveOverlayLayers.swift`: reusable Core Animation layer subclasses for live selection flow and
   scrim masking.
 - `LiveChromePlacement.swift`: live HUD/loupe text metrics, pending color text, and deterministic
