+++ native/macos-host/Sources/RsnapNativeHostKit/LiveOverlayRenderer.swift
@@ -32,16 +32,6 @@ struct LivePreviewSnapshot {
 }
 
 @MainActor
-private enum LiveOverlayTypography {
-	static let font = NSFont.monospacedSystemFont(ofSize: 13, weight: .medium)
-	static let lineHeight = ceil("x=0".size(using: font).height)
-	static let commaWidth = ",".size(using: font).width
-	static let keycapTextSize = "Tab".size(using: font)
-	static let keycapFrameSize = CGSize(
-		width: keycapTextSize.width + 12, height: keycapTextSize.height + 4)
-}
-
-@MainActor
 final class LiveOverlayRenderer {
 	private weak var hostView: NSView?
 	private let rootLayer = CALayer()
@@ -95,19 +85,6 @@ final class LiveOverlayRenderer {
 		category: "LiveChromeTelemetry"
 	)
 	private static let activeInputWindow: TimeInterval = 0.25
-	private static let hudColorResolveAnimationKey = "rsnap.hud.color.resolve"
-	private static let hudColorResolveBackgroundAnimationKey = "rsnap.hud.color.resolve.background"
-	private static let hudColorRollAnimationKey = "rsnap.hud.color.roll"
-	private static let hudColorPendingRollAnimationKey = "rsnap.hud.color.pending.roll"
-	private static let hudColorRollDuration: TimeInterval = 0.40
-	private static let hudColorRollDigitStagger: TimeInterval = 0.024
-	private static let hexWheel = Array("0123456789ABCDEF")
-	private static let pendingHexRollBaseSeed: UInt64 = 0x5EED_71A5_C01D
-	private struct HudHexPendingRollColumnState {
-		let digits: [Character]
-		let scrollsUp: Bool
-		let contentLayer: CALayer
-	}
 	private enum LayerZ {
 		static let root: CGFloat = 100
 		static let chromeRoot: CGFloat = 300
@@ -123,16 +100,14 @@ final class LiveOverlayRenderer {
 	private var lastRenderedFocusFlowAnimates = false
 	private var lastChromeRenderUptime: TimeInterval?
 	private var lastActiveChromeRenderUptime: TimeInterval?
-	private var lastHudColorPending: Bool?
-	private var hudColorRevealArmed = true
-	private var hasResolvedHudColor = false
-	private var lastResolvedHudHexText: String?
-	private var lastResolvedHudSwatchColor: CGColor?
-	private var activeHudHexRollTarget: String?
-	private var activeHudHexRollSwatchColor: CGColor?
-	private var hudHexRollAnimationEndUptime: TimeInterval?
-	private var hudHexPendingRollActive = false
-	private var hudHexPendingRollColumns: [HudHexPendingRollColumnState] = []
+	private lazy var hudColorRollCoordinator = LiveHudColorRollCoordinator(
+		hudHexLayer: hudHexLayer,
+		hudHexRollLayer: hudHexRollLayer,
+		hudSwatchLayer: hudSwatchLayer,
+		backingScaleProvider: { [weak self] in
+			self?.hostView?.window?.backingScaleFactor ?? 2
+		}
+	)
 
 	init(hostView: NSView) {
 		self.hostView = hostView
@@ -323,7 +298,7 @@ final class LiveOverlayRenderer {
 		lastRenderedFocusFlowAnimates = false
 		lastChromeRenderUptime = nil
 		lastActiveChromeRenderUptime = nil
-		resetHudColorAnimationState()
+		hudColorRollCoordinator.reset()
 		hoverFlowLayer.hide()
 	}
 
@@ -691,7 +666,7 @@ final class LiveOverlayRenderer {
 	private func renderHud(_ snapshot: LivePreviewSnapshot) {
 		guard let hudFrame = snapshot.hudFrame else {
 			hudLayer.isHidden = true
-			resetHudColorAnimationState()
+			hudColorRollCoordinator.reset()
 			return
 		}
 		let palette = CaptureChrome.palette(for: snapshot.theme, settings: snapshot.settings)
@@ -732,42 +707,24 @@ final class LiveOverlayRenderer {
 		)
 		cursorX += positionSize.width + CaptureChrome.hudGroupSpacing
 
-		hudSwatchLayer.frame = CGRect(
+		let swatchFrame = CGRect(
 			x: cursorX,
 			y: hudLayer.bounds.midY - swatchSize.height / 2,
 			width: swatchSize.width,
 			height: swatchSize.height
 		)
-		hudSwatchLayer.cornerRadius = 0
-		let pendingSwatchColor = palette.labelText.withAlphaComponent(0.16)
-		let swatchColor =
-			snapshot.rgbSample.map {
-				NSColor(
-					calibratedRed: CGFloat($0.r) / 255, green: CGFloat($0.g) / 255,
-					blue: CGFloat($0.b) / 255, alpha: 1)
-			} ?? pendingSwatchColor
-		hudSwatchLayer.backgroundColor = swatchColor.cgColor
-		hudSwatchLayer.borderColor = palette.swatchStroke.cgColor
-		hudSwatchLayer.borderWidth = 1
 		cursorX += swatchSize.width + CaptureChrome.hudColorItemSpacing
 
-		let hexTextColor =
-			snapshot.colorDisplay.isPending
-			? palette.labelText.withAlphaComponent(0.46) : palette.labelText
 		let hexFrame = CGRect(
 			x: cursorX, y: baselineY, width: ceil(snapshot.colorDisplay.hexSlotWidth),
 			height: ceil(LiveOverlayTypography.lineHeight))
-		applyText(
-			hudHexLayer, text: snapshot.colorDisplay.hexText, font: font, color: hexTextColor,
-			frame: hexFrame, alignment: .left)
-		updateHudColorAnimation(
-			isPending: snapshot.colorDisplay.isPending,
-			pendingSwatchColor: pendingSwatchColor,
-			resolvedSwatchColor: swatchColor,
-			resolvedHexText: snapshot.colorDisplay.hexText,
+		hudColorRollCoordinator.render(
+			colorDisplay: snapshot.colorDisplay,
+			rgbSample: snapshot.rgbSample,
+			palette: palette,
+			swatchFrame: swatchFrame,
 			hexFrame: hexFrame,
-			font: font,
-			textColor: palette.labelText
+			font: font
 		)
 		cursorX += snapshot.colorDisplay.hexSlotWidth + CaptureChrome.hudGroupSpacing
 
@@ -795,702 +752,6 @@ final class LiveOverlayRenderer {
 			hudKeycapLayer.isHidden = true
 			hudKeycapTextLayer.isHidden = true
 		}
-	}
-
-	private func updateHudColorAnimation(
-		isPending: Bool,
-		pendingSwatchColor: NSColor,
-		resolvedSwatchColor: NSColor,
-		resolvedHexText: String,
-		hexFrame: CGRect,
-		font: NSFont,
-		textColor: NSColor
-	) {
-		if isPending {
-			hudSwatchLayer.removeAnimation(forKey: Self.hudColorResolveAnimationKey)
-			hudSwatchLayer.removeAnimation(forKey: Self.hudColorResolveBackgroundAnimationKey)
-			hudSwatchLayer.opacity = 1
-			hudHexLayer.removeAnimation(forKey: Self.hudColorResolveAnimationKey)
-			if hasResolvedHudColor {
-				clearHudHexRollAnimation()
-				if let lastResolvedHudSwatchColor {
-					hudSwatchLayer.backgroundColor = lastResolvedHudSwatchColor
-				}
-				if let lastResolvedHudHexText {
-					applyText(
-						hudHexLayer,
-						text: lastResolvedHudHexText,
-						font: font,
-						color: textColor,
-						frame: hexFrame,
-						alignment: .left
-					)
-				}
-				hudHexLayer.isHidden = false
-				lastHudColorPending = false
-				hudColorRevealArmed = false
-				return
-			}
-			beginOrUpdateHudHexPendingRollAnimation(
-				frame: hexFrame,
-				font: font,
-				textColor: textColor
-			)
-			lastHudColorPending = true
-			return
-		}
-
-		let wasPending = lastHudColorPending == true
-		let shouldAnimateReveal = wasPending && hudColorRevealArmed && !hasResolvedHudColor
-		let priorSwatchColor =
-			wasPending ? hudSwatchLayer.presentation()?.backgroundColor : nil
-		let priorSwatchOpacity =
-			wasPending ? hudSwatchLayer.presentation()?.opacity : nil
-		let priorHexOpacity = wasPending ? hudHexLayer.presentation()?.opacity : nil
-		lastResolvedHudHexText = resolvedHexText
-		lastResolvedHudSwatchColor = resolvedSwatchColor.cgColor
-		hasResolvedHudColor = true
-		lastHudColorPending = false
-		hudColorRevealArmed = false
-
-		guard shouldAnimateReveal else {
-			updateHudHexRollVisibility(
-				target: resolvedHexText,
-				frame: hexFrame,
-				font: font,
-				textColor: textColor
-			)
-			return
-		}
-
-		addHudColorResolveAnimation(
-			to: hudSwatchLayer,
-			fromOpacity: priorSwatchOpacity.map(CGFloat.init) ?? 0.62
-		)
-		beginHudHexRollAnimation(
-			target: resolvedHexText,
-			frame: hexFrame,
-			font: font,
-			textColor: textColor,
-			initialOpacity: priorHexOpacity.map(CGFloat.init) ?? 0.62,
-			targetSwatchColor: resolvedSwatchColor
-		)
-		let colorAnimation = CABasicAnimation(keyPath: "backgroundColor")
-		colorAnimation.fromValue = priorSwatchColor ?? pendingSwatchColor.cgColor
-		colorAnimation.toValue = resolvedSwatchColor.cgColor
-		colorAnimation.duration = 0.16
-		colorAnimation.timingFunction = CAMediaTimingFunction(name: .easeOut)
-		hudSwatchLayer.add(
-			colorAnimation,
-			forKey: Self.hudColorResolveBackgroundAnimationKey
-		)
-	}
-
-	private func addHudColorResolveAnimation(to layer: CALayer, fromOpacity: CGFloat) {
-		let animation = CABasicAnimation(keyPath: "opacity")
-		animation.fromValue = fromOpacity
-		animation.toValue = 1
-		animation.duration = 0.16
-		animation.timingFunction = CAMediaTimingFunction(name: .easeOut)
-		layer.add(animation, forKey: Self.hudColorResolveAnimationKey)
-	}
-
-	private func updateHudHexRollVisibility(
-		target: String,
-		frame: CGRect,
-		font: NSFont,
-		textColor: NSColor
-	) {
-		guard let activeTarget = activeHudHexRollTarget else {
-			clearHudHexRollAnimation()
-			hudHexLayer.isHidden = false
-			return
-		}
-		let now = ProcessInfo.processInfo.systemUptime
-		if activeTarget != target {
-			if let animationEnd = hudHexRollAnimationEndUptime {
-				if now < animationEnd {
-					if let activeHudHexRollSwatchColor {
-						hudSwatchLayer.backgroundColor = activeHudHexRollSwatchColor
-					}
-					hudHexLayer.isHidden = true
-					hudHexRollLayer.isHidden = false
-					hudHexRollLayer.frame = frame
-					return
-				}
-				finishHudHexRollAnimation()
-			}
-			clearHudHexRollAnimation()
-			hudHexLayer.isHidden = false
-			return
-		}
-		if let animationEnd = hudHexRollAnimationEndUptime,
-			now >= animationEnd
-		{
-			finishHudHexRollAnimation()
-		}
-
-		hudHexLayer.isHidden = true
-		hudHexRollLayer.isHidden = false
-		hudHexRollLayer.frame = frame
-	}
-
-	private func finishHudHexRollAnimation() {
-		hudHexRollAnimationEndUptime = nil
-		activeHudHexRollSwatchColor = nil
-		hudHexPendingRollActive = false
-		hudHexPendingRollColumns.removeAll(keepingCapacity: true)
-		removeHudHexRollLayerAnimations()
-	}
-
-	private func beginOrUpdateHudHexPendingRollAnimation(
-		frame: CGRect,
-		font: NSFont,
-		textColor: NSColor
-	) {
-		hudHexLayer.isHidden = true
-		hudHexRollLayer.isHidden = false
-		hudHexRollLayer.frame = frame
-		guard hudHexPendingRollActive == false else {
-			return
-		}
-
-		clearHudHexRollAnimation()
-		hudHexPendingRollActive = true
-		hudHexPendingRollColumns.removeAll(keepingCapacity: true)
-		hudHexLayer.isHidden = true
-		hudHexRollLayer.isHidden = false
-		hudHexRollLayer.frame = frame
-
-		let lineHeight = ceil(LiveOverlayTypography.lineHeight)
-		let characterFrames = hudHexCharacterFrames(
-			for: "#FFFFFF",
-			font: font,
-			lineHeight: lineHeight
-		)
-		let hashLayer = makeHudHexRollTextLayer(
-			text: "#",
-			font: font,
-			color: textColor.withAlphaComponent(0.72),
-			frame: characterFrames.first ?? CGRect(x: 0, y: 0, width: 0, height: lineHeight)
-		)
-		hudHexRollLayer.addSublayer(hashLayer)
-
-		for index in 0..<6 {
-			let characterFrame =
-				index + 1 < characterFrames.count
-				? characterFrames[index + 1]
-				: CGRect(x: 0, y: 0, width: 0, height: lineHeight)
-			let columnLayer = CALayer()
-			columnLayer.masksToBounds = true
-			columnLayer.frame = characterFrame
-			hudHexRollLayer.addSublayer(columnLayer)
-			let columnState = addHudHexPendingRollColumn(
-				to: columnLayer,
-				index: index,
-				font: font,
-				textColor: textColor,
-				lineHeight: lineHeight,
-				digitWidth: characterFrame.width
-			)
-			hudHexPendingRollColumns.append(columnState)
-		}
-	}
-	private static func pendingHexRollSeed(index: Int) -> UInt64 {
-		let uptimeBucket = UInt64((ProcessInfo.processInfo.systemUptime * 1_000).rounded(.down))
-		let mixedIndex = UInt64(index + 1) &* 0x9E37_79B9_7F4A_7C15
-		return pendingHexRollBaseSeed ^ uptimeBucket ^ mixedIndex
-	}
-
-	private static func pendingHexRollSequence(index: Int) -> [Character] {
-		var seed = pendingHexRollSeed(index: index)
-		seed = seed &* 6_364_136_223_846_793_005 &+ 1_442_695_040_888_963_407
-		let visibleRows = 47 + Int((seed >> 57) & 0x1F) + index * 3
-		var digits: [Character] = []
-		digits.reserveCapacity(visibleRows + 1)
-		var previous: Character?
-		for offset in 0..<visibleRows {
-			seed = seed &* 6_364_136_223_846_793_005 &+ 1_442_695_040_888_963_407
-			var wheelIndex = Int((seed >> 58) & 0xF)
-			if let previousDigit = previous,
-				Self.hexWheel[wheelIndex] == previousDigit
-			{
-				wheelIndex = (wheelIndex + offset + index + 1) % Self.hexWheel.count
-			}
-			let digit = Self.hexWheel[wheelIndex]
-			digits.append(digit)
-			previous = digit
-		}
-		if let first = digits.first {
-			digits.append(first)
-		}
-		return digits
-	}
-
-	private static func pendingHexRollColumnDuration(index: Int) -> TimeInterval {
-		let seed =
-			pendingHexRollSeed(index: index)
-			&* 2_862_933_555_777_941_757
-			&+ 3_037_000_493
-		return 1.58 + Double((seed >> 56) & 0x1F) * 0.031
-	}
-
-	private static func pendingHexRollColumnPhase(index: Int, duration: TimeInterval)
-		-> TimeInterval
-	{
-		let seed =
-			pendingHexRollSeed(index: index)
-			&* 11_400_714_819_323_198_485
-			&+ 12_829_314
-		let ratio = Double((seed >> 40) & 0xFFFF) / 65_535.0
-		return duration * ratio
-	}
-
-	private static func pendingHexRollColumnScrollsUp(index: Int) -> Bool {
-		let uptimeBucket = UInt64((ProcessInfo.processInfo.systemUptime * 1_000).rounded(.down))
-		let startsUp = ((pendingHexRollBaseSeed ^ uptimeBucket) & 1) == 0
-		if index <= 1 {
-			return index == 0 ? startsUp : !startsUp
-		}
-		let seed =
-			pendingHexRollSeed(index: index)
-			&* 3_202_034_522_624_059_733
-			&+ 1_029
-		return ((seed >> 63) & 1) == 0
-	}
-
-	private static func resolveHexRollColumnScrollsUp(
-		index: Int,
-		startDigit: Character,
-		targetDigit: Character
-	) -> Bool {
-		let startValue = UInt64(startDigit.unicodeScalars.first?.value ?? 0)
-		let targetValue = UInt64(targetDigit.unicodeScalars.first?.value ?? 0)
-		let seed =
-			pendingHexRollSeed(index: index)
-			^ (startValue &* 1_099_511_628_211)
-			^ (targetValue &* 2_862_933_555_777_941_757)
-		let startsUp = ((seed >> 63) & 1) == 0
-		if index <= 1 {
-			return index == 0 ? startsUp : !startsUp
-		}
-		return ((seed >> 59) & 1) == 0
-	}
-
-	private static func resolveHexRollExtraLoops(index: Int, targetDigit: Character) -> Int {
-		let targetValue = UInt64(targetDigit.unicodeScalars.first?.value ?? 0)
-		let seed =
-			pendingHexRollSeed(index: index)
-			^ (targetValue &* 11_400_714_819_323_198_485)
-		return 1 + Int((seed >> 60) & 1)
-	}
-
-	private static func resolveHexRollSequence(
-		from startDigit: Character,
-		to targetDigit: Character,
-		index: Int,
-		scrollsUp: Bool
-	) -> [Character] {
-		let wheelCount = max(hexWheel.count, 1)
-		let startIndex = hexWheel.firstIndex(of: startDigit) ?? 0
-		let targetIndex = hexWheel.firstIndex(of: targetDigit) ?? startIndex
-		let directedDistance =
-			scrollsUp
-			? (targetIndex - startIndex + wheelCount) % wheelCount
-			: (startIndex - targetIndex + wheelCount) % wheelCount
-		let extraSteps =
-			resolveHexRollExtraLoops(index: index, targetDigit: targetDigit)
-			* wheelCount
-		let totalSteps =
-			directedDistance + extraSteps
-		return (0...totalSteps).map { offset in
-			let wheelIndex =
-				scrollsUp
-				? (startIndex + offset) % wheelCount
-				: (startIndex - offset + (totalSteps + wheelCount) * wheelCount) % wheelCount
-			return hexWheel[wheelIndex]
-		}
-	}
-
-	private func beginHudHexRollAnimation(
-		target: String,
-		frame: CGRect,
-		font: NSFont,
-		textColor: NSColor,
-		initialOpacity: CGFloat,
-		targetSwatchColor: NSColor? = nil
-	) {
-		let lineHeight = ceil(LiveOverlayTypography.lineHeight)
-		let startDigits = currentPendingHudHexDigits(lineHeight: lineHeight)
-		let pendingDirections = hudHexPendingRollColumns.map(\.scrollsUp)
-		clearHudHexRollAnimation()
-		activeHudHexRollTarget = target
-		activeHudHexRollSwatchColor = targetSwatchColor?.cgColor
-		let now = ProcessInfo.processInfo.systemUptime
-		let targetDigits = Array(target.dropFirst())
-		var rollEndOffset: TimeInterval = 0
-		hudHexLayer.isHidden = true
-		hudHexRollLayer.isHidden = false
-		hudHexRollLayer.frame = frame
-
-		let characterFrames = hudHexCharacterFrames(
-			for: target,
-			font: font,
-			lineHeight: lineHeight
-		)
-		let hashLayer = makeHudHexRollTextLayer(
-			text: "#",
-			font: font,
-			color: textColor.withAlphaComponent(0.72),
-			frame: characterFrames.first ?? CGRect(x: 0, y: 0, width: 0, height: lineHeight)
-		)
-		hudHexRollLayer.addSublayer(hashLayer)
-
-		for (index, targetDigit) in targetDigits.enumerated() {
-			let characterFrame =
-				index + 1 < characterFrames.count
-				? characterFrames[index + 1]
-				: CGRect(x: 0, y: 0, width: 0, height: lineHeight)
-			let columnLayer = CALayer()
-			columnLayer.masksToBounds = true
-			columnLayer.frame = characterFrame
-			hudHexRollLayer.addSublayer(columnLayer)
-			let startDigit =
-				index < startDigits.count
-				? startDigits[index]
-				: nil
-			let resolvedStartDigit = startDigit ?? Self.hexWheel.first ?? targetDigit
-			let scrollsUp =
-				index < pendingDirections.count
-				? pendingDirections[index]
-				: Self.resolveHexRollColumnScrollsUp(
-					index: index,
-					startDigit: resolvedStartDigit,
-					targetDigit: targetDigit
-				)
-			let columnEndOffset = addHudHexRollDigit(
-				to: columnLayer,
-				startDigit: resolvedStartDigit,
-				targetDigit: targetDigit,
-				index: index,
-				font: font,
-				textColor: textColor,
-				initialOpacity: initialOpacity,
-				lineHeight: lineHeight,
-				digitWidth: characterFrame.width,
-				scrollsUp: scrollsUp
-			)
-			hudHexPendingRollColumns.append(columnEndOffset.state)
-			rollEndOffset = max(rollEndOffset, columnEndOffset.endOffset)
-		}
-		hudHexRollAnimationEndUptime = now + rollEndOffset + 0.03
-	}
-
-	private func hudHexCharacterFrames(
-		for text: String,
-		font: NSFont,
-		lineHeight: CGFloat
-	) -> [CGRect] {
-		let characters = Array(text)
-		return characters.indices.map { index in
-			let prefixStart = String(characters.prefix(index)).size(using: font).width
-			let prefixEnd = String(characters.prefix(index + 1)).size(using: font).width
-			return CGRect(
-				x: prefixStart,
-				y: 0,
-				width: max(prefixEnd - prefixStart, 1),
-				height: lineHeight
-			)
-		}
-	}
-
-	private func addHudHexPendingRollColumn(
-		to columnLayer: CALayer,
-		index: Int,
-		font: NSFont,
-		textColor: NSColor,
-		lineHeight: CGFloat,
-		digitWidth: CGFloat
-	) -> HudHexPendingRollColumnState {
-		var digits = Self.pendingHexRollSequence(index: index)
-		let scrollsUp = Self.pendingHexRollColumnScrollsUp(index: index)
-		if scrollsUp == false {
-			digits.reverse()
-		}
-		let contentText = digits.map(String.init).joined(separator: "\n")
-		let contentLayer = CALayer()
-		contentLayer.frame = CGRect(
-			x: 0,
-			y: 0,
-			width: digitWidth,
-			height: lineHeight * CGFloat(digits.count)
-		)
-		columnLayer.addSublayer(contentLayer)
-
-		let digitLayer = makeHudHexRollMultilineTextLayer(
-			text: contentText,
-			font: font,
-			color: textColor.withAlphaComponent(0.72),
-			lineHeight: lineHeight,
-			frame: contentLayer.bounds
-		)
-		contentLayer.addSublayer(digitLayer)
-
-		let animation = CABasicAnimation(keyPath: "transform.translation.y")
-		let travel = lineHeight * CGFloat(max(digits.count - 1, 1))
-		animation.fromValue = scrollsUp ? 0 : -travel
-		animation.toValue = scrollsUp ? -travel : 0
-		let duration = Self.pendingHexRollColumnDuration(index: index)
-		animation.duration = duration
-		animation.beginTime =
-			CACurrentMediaTime()
-			- Self.pendingHexRollColumnPhase(index: index, duration: duration)
-		animation.repeatCount = .infinity
-		animation.timingFunction = CAMediaTimingFunction(name: .linear)
-		animation.isRemovedOnCompletion = false
-		contentLayer.add(animation, forKey: Self.hudColorPendingRollAnimationKey)
-		return HudHexPendingRollColumnState(
-			digits: digits,
-			scrollsUp: scrollsUp,
-			contentLayer: contentLayer
-		)
-	}
-
-	private func currentPendingHudHexDigits(lineHeight: CGFloat) -> [Character?] {
-		hudHexPendingRollColumns.map { column in
-			guard column.digits.isEmpty == false else {
-				return nil
-			}
-			let presentationLayer = column.contentLayer.presentation() ?? column.contentLayer
-			let translationY = presentationLayer.transform.m42
-			let rawIndex = Int((-translationY / lineHeight).rounded())
-			let visibleIndex = min(max(rawIndex, 0), column.digits.count - 1)
-			return column.digits[visibleIndex]
-		}
-	}
-
-	private func addHudHexRollDigit(
-		to columnLayer: CALayer,
-		startDigit: Character,
-		targetDigit: Character,
-		index: Int,
-		font: NSFont,
-		textColor: NSColor,
-		initialOpacity: CGFloat,
-		lineHeight: CGFloat,
-		digitWidth: CGFloat,
-		scrollsUp: Bool
-	) -> (state: HudHexPendingRollColumnState, endOffset: TimeInterval) {
-		let rollDigits = Self.resolveHexRollSequence(
-			from: startDigit,
-			to: targetDigit,
-			index: index,
-			scrollsUp: scrollsUp
-		)
-		let terminalPaddingRows = 2
-		let contentDigits: [Character]
-		let startRowIndex: Int
-		let targetRowIndex: Int
-		if scrollsUp {
-			contentDigits =
-				rollDigits
-				+ Array(repeating: targetDigit, count: terminalPaddingRows)
-			startRowIndex = 0
-			targetRowIndex = max(rollDigits.count - 1, 0)
-		} else {
-			contentDigits =
-				Array(repeating: targetDigit, count: terminalPaddingRows)
-				+ Array(rollDigits.reversed())
-			startRowIndex = max(contentDigits.count - 1, 0)
-			targetRowIndex = terminalPaddingRows
-		}
-		let contentLayer = CALayer()
-		contentLayer.opacity = Float(max(initialOpacity, 0.72))
-		contentLayer.frame = CGRect(
-			x: 0,
-			y: 0,
-			width: digitWidth,
-			height: lineHeight * CGFloat(contentDigits.count)
-		)
-		columnLayer.addSublayer(contentLayer)
-
-		addHudHexRollDigitStack(
-			to: contentLayer,
-			digits: contentDigits,
-			font: font,
-			color: textColor,
-			lineHeight: lineHeight,
-			digitWidth: digitWidth
-		)
-
-		let fromY = -lineHeight * CGFloat(startRowIndex)
-		let toY = -lineHeight * CGFloat(targetRowIndex)
-		contentLayer.transform = CATransform3DMakeTranslation(0, toY, 0)
-
-		let stagger = Double(index) * Self.hudColorRollDigitStagger
-		let duration =
-			Self.hudColorRollDuration
-			+ Double(Self.resolveHexRollExtraLoops(index: index, targetDigit: targetDigit)) * 0.035
-		let animation = CABasicAnimation(keyPath: "transform.translation.y")
-		animation.fromValue = fromY
-		animation.toValue = toY
-		animation.beginTime = CACurrentMediaTime() + stagger
-		animation.duration = duration
-		animation.timingFunction = CAMediaTimingFunction(name: .easeOut)
-		animation.fillMode = .both
-		animation.isRemovedOnCompletion = false
-		contentLayer.add(animation, forKey: Self.hudColorRollAnimationKey)
-		let columnState = HudHexPendingRollColumnState(
-			digits: contentDigits,
-			scrollsUp: scrollsUp,
-			contentLayer: contentLayer
-		)
-		return (columnState, stagger + duration)
-	}
-
-	private func addHudHexRollDigitStack(
-		to contentLayer: CALayer,
-		digits: [Character],
-		font: NSFont,
-		color: NSColor,
-		lineHeight: CGFloat,
-		digitWidth: CGFloat
-	) {
-		for (row, digit) in digits.enumerated() {
-			let digitLayer = makeHudHexRollTextLayer(
-				text: String(digit),
-				font: font,
-				color: color,
-				frame: CGRect(
-					x: 0,
-					y: CGFloat(row) * lineHeight,
-					width: digitWidth,
-					height: lineHeight
-				)
-			)
-			contentLayer.addSublayer(digitLayer)
-		}
-	}
-
-	private func removeHudHexRollLayerAnimations() {
-		hudHexRollLayer.removeAllAnimations()
-		for sublayer in hudHexRollLayer.sublayers ?? [] {
-			removeAnimationsRecursively(from: sublayer)
-		}
-	}
-
-	private func removeAnimationsRecursively(from layer: CALayer) {
-		layer.removeAllAnimations()
-		for sublayer in layer.sublayers ?? [] {
-			removeAnimationsRecursively(from: sublayer)
-		}
-	}
-
-	private func makeHudHexRollTextLayer(
-		text: String,
-		font: NSFont,
-		color: NSColor,
-		frame: CGRect
-	) -> CATextLayer {
-		let layer = CATextLayer()
-		layer.contentsScale = hostView?.window?.backingScaleFactor ?? 2
-		layer.string = text
-		layer.font = font
-		layer.fontSize = font.pointSize
-		layer.foregroundColor = color.cgColor
-		layer.alignmentMode = .left
-		layer.frame = frame
-		layer.isWrapped = false
-		return layer
-	}
-
-	private func makeHudHexRollMultilineTextLayer(
-		text: String,
-		font: NSFont,
-		color: NSColor,
-		lineHeight: CGFloat,
-		frame: CGRect
-	) -> CATextLayer {
-		let paragraphStyle = NSMutableParagraphStyle()
-		paragraphStyle.alignment = .left
-		paragraphStyle.lineBreakMode = .byClipping
-		paragraphStyle.minimumLineHeight = lineHeight
-		paragraphStyle.maximumLineHeight = lineHeight
-		let attributedString = NSAttributedString(
-			string: text,
-			attributes: [
-				.font: font,
-				.foregroundColor: color,
-				.paragraphStyle: paragraphStyle,
-			]
-		)
-		let layer = CATextLayer()
-		layer.contentsScale = hostView?.window?.backingScaleFactor ?? 2
-		layer.string = attributedString
-		layer.alignmentMode = .left
-		layer.frame = frame
-		layer.isWrapped = true
-		layer.truncationMode = .none
-		return layer
-	}
-
-	private func addHudRollAnimation(
-		to layer: CALayer,
-		fromY: CGFloat,
-		toY: CGFloat,
-		opacityValues: [CGFloat],
-		keyTimes: [CGFloat],
-		beginOffset: TimeInterval,
-		duration: TimeInterval,
-		timing: CAMediaTimingFunctionName = .easeInEaseOut
-	) {
-		let translation = CABasicAnimation(keyPath: "transform.translation.y")
-		translation.fromValue = fromY
-		translation.toValue = toY
-
-		let opacity = CAKeyframeAnimation(keyPath: "opacity")
-		opacity.values = opacityValues.map { NSNumber(value: Double($0)) }
-		opacity.keyTimes = keyTimes.map { NSNumber(value: Double($0)) }
-
-		let group = CAAnimationGroup()
-		group.animations = [translation, opacity]
-		group.beginTime = CACurrentMediaTime() + beginOffset
-		group.duration = duration
-		group.timingFunction = CAMediaTimingFunction(name: timing)
-		group.fillMode = .both
-		group.isRemovedOnCompletion = false
-		layer.add(group, forKey: Self.hudColorRollAnimationKey)
-	}
-
-	private func clearHudHexRollAnimation() {
-		activeHudHexRollTarget = nil
-		activeHudHexRollSwatchColor = nil
-		hudHexRollAnimationEndUptime = nil
-		hudHexPendingRollActive = false
-		hudHexPendingRollColumns.removeAll(keepingCapacity: true)
-		removeHudHexRollLayerAnimations()
-		for sublayer in hudHexRollLayer.sublayers ?? [] {
-			sublayer.removeFromSuperlayer()
-		}
-		hudHexRollLayer.isHidden = true
-	}
-
-	private func resetHudColorAnimationState() {
-		lastHudColorPending = nil
-		hudColorRevealArmed = true
-		hasResolvedHudColor = false
-		lastResolvedHudHexText = nil
-		lastResolvedHudSwatchColor = nil
-		activeHudHexRollTarget = nil
-		activeHudHexRollSwatchColor = nil
-		hudHexRollAnimationEndUptime = nil
-		hudHexPendingRollActive = false
-		hudHexPendingRollColumns.removeAll(keepingCapacity: true)
-		hudSwatchLayer.removeAnimation(forKey: Self.hudColorResolveAnimationKey)
-		hudSwatchLayer.removeAnimation(forKey: Self.hudColorResolveBackgroundAnimationKey)
-		hudHexLayer.removeAnimation(forKey: Self.hudColorResolveAnimationKey)
-		hudHexLayer.isHidden = false
-		clearHudHexRollAnimation()
 	}
 
 	private func renderLoupe(_ snapshot: LivePreviewSnapshot) {
