+++ native/macos-host/Sources/RsnapNativeHostKit/LiveOverlayTypography.swift
@@ -1,0 +1,12 @@
+import AppKit
+import CoreGraphics
+
+@MainActor
+enum LiveOverlayTypography {
+	static let font = NSFont.monospacedSystemFont(ofSize: 13, weight: .medium)
+	static let lineHeight = ceil("x=0".size(using: font).height)
+	static let commaWidth = ",".size(using: font).width
+	static let keycapTextSize = "Tab".size(using: font)
+	static let keycapFrameSize = CGSize(
+		width: keycapTextSize.width + 12, height: keycapTextSize.height + 4)
+}
